feat: refact plugins dir
