+++ packages/vite-service/src/plugins/index.ts
@@ -1,0 +1,5 @@
+export * from './external';
+export * from './history';
+export * from './import';
+export * from './polyfill';
+export * from './spa';
